refactor(echopay): move ClientInterface to Souk as HttpClientInterface
The HTTP client contract is generic (get/post/patch/delete returning arrays)
and not EchoPay-specific, so it belongs in the shared payments layer rather
than the connector. Renamed to HttpClientInterface to disambiguate at the
Souk level and placed alongside the other payment contracts. Client,
MockClient, and EchoPayService updated to the new namespace; the now-empty
EchoPay/Contracts directory is removed.
